Some fixes and architecture improvements Resolved some tests troubles and ProfileEventsViewSet's get_queryset method; Separated viewsets in every app; Added debug_toolbar to requirements/development.txt
