+++ apps/chats/viewsets/chats.py
@@ -1,8 +1,5 @@
-import logging
-
 from asgiref.sync import async_to_sync
 from django.contrib.auth import get_user_model
-from django.db.models import Q
 from django.forms import model_to_dict
 from drf_yasg.utils import swagger_auto_schema, no_body
 from rest_framework import viewsets, status
@@ -16,7 +13,6 @@ from apps.profiles.models import User
 from apps.chats.managers import ChatManager
 from apps.chats.models import Chat, Message
 from apps.chats.permissions.chats import ChatPermissions, DirectChatPermissions
-from apps.chats.permissions.message import MessagePermissions
 from apps.chats.serializers.chats import (
     ChatRetrieveSerializer,
     ChatListSerializer,
@@ -26,8 +22,6 @@ from apps.chats.serializers.messages import (
     MessageCreateSerializer,
 )
 from apps.chats.utils import list_chats_raw
-
-logger = logging.getLogger("chat_app")
 
 user_model = get_user_model()
 
@@ -127,36 +121,6 @@ class ChatViewSet(viewsets.ModelViewSet):
     @swagger_auto_schema(auto_schema=None)
     def create(self, request, *args, **kwargs):
         pass
-
-
-class MessageViewSet(viewsets.ModelViewSet):
-    model = Message
-    permission_classes = [IsAuthenticated, MessagePermissions]
-    lookup_url_kwarg = "message_id"
-    http_method_names = ["get", "patch", "delete"]
-    queryset = model.objects.all()
-
-    def get_serializer_class(self):
-        match self.action:
-            case "retrieve":
-                return MessageRetrieveSerializer
-            case "list":
-                return MessageRetrieveSerializer
-            case "partial_update":
-                return MessageCreateSerializer
-            case _:
-                return EmptySerializer
-
-    def get_queryset(self):
-        if self.kwargs.get("message_id"):
-            self.queryset = self.model.objects.all()
-        else:
-            self.queryset = self.model.objects.filter(
-                id__in=Chat.objects.filter(
-                    participants__id=self.request.user.id)
-                .values_list("chat_messages", flat=True)
-            ).order_by("-created_at")
-        return self.queryset.all()
 
 
 class DirectChatViewSet(viewsets.ModelViewSet):
