+++ apps/core/urls.py
@@ -3,17 +3,21 @@ from drf_yasg import openapi
 from drf_yasg.views import get_schema_view
 from rest_framework import permissions
 
-from .views import (
+from .views.tokens import (
     DecoratedTokenObtainPairView,
     DecoratedTokenBlacklistView,
     DecoratedTokenVerifyView,
     DecoratedTokenRefreshView,
-    RegisterView,
-    VerifyEmailView,
-    ReverifyEmailView,
+)
+from .views.register import RegisterView
+from .views.passwords import (
     PasswordResetView,
     PasswordResetConfirmView,
     PasswordChangeView,
+)
+from .views.emails import (
+    VerifyEmailView,
+    ReverifyEmailView,
     CheckEmailExistsView,
     ChangeEmailView,
 )
