+++ apps/events/tests/events/tests_created_by.py
@@ -119,11 +119,11 @@ def test_event_created_by_declined_private(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_without_following(
+def test_event_private_created_by_without_following(
         api_client,
         user,
         user_2,
-        event_not_visible_created_by_user_2,
+        event_private_created_by_user_2,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -135,12 +135,12 @@ def test_event_not_visible_created_by_without_following(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_accepted(
+def test_event_private_created_by_accepted(
         api_client,
         user,
         user_2,
         follower_user_accepted,
-        event_not_visible_created_by_user_2,
+        event_private_created_by_user_2,
 ):
     token = get_tokens(user)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -152,12 +152,12 @@ def test_event_not_visible_created_by_accepted(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_accepted_private(
+def test_event_private_created_by_accepted_private(
         api_client,
         user_private,
         user_2_private,
         follower_user_accepted_private,
-        event_not_visible_created_by_user_2_private,
+        event_private_created_by_user_2_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -169,12 +169,12 @@ def test_event_not_visible_created_by_accepted_private(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_pending_private(
+def test_event_private_created_by_pending_private(
         api_client,
         user_private,
         user_2_private,
         follower_user_pending_private,
-        event_not_visible_created_by_user_2_private,
+        event_private_created_by_user_2_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -186,12 +186,12 @@ def test_event_not_visible_created_by_pending_private(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_declined_private(
+def test_event_private_created_by_declined_private(
         api_client,
         user_private,
         user_2_private,
         follower_user_declined_private,
-        event_not_visible_created_by_user_2_private,
+        event_private_created_by_user_2_private,
 ):
     token = get_tokens(user_private)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -219,10 +219,10 @@ def test_event_created_by_current_user(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_current_user(
+def test_event_private_created_by_current_user(
         api_client,
         user_2,
-        event_not_visible_created_by_user_2,
+        event_private_created_by_user_2,
 ):
     token = get_tokens(user_2)
     api_client.credentials(HTTP_AUTHORIZATION="Bearer " + token)
@@ -231,7 +231,7 @@ def test_event_not_visible_created_by_current_user(
     )
     assert response.status_code == 200
     assert response.data[0].get("id") is not None
-    assert response.data[0].get("name") == event_not_visible_created_by_user_2.name
+    assert response.data[0].get("name") == event_private_created_by_user_2.name
 
 
 @pytest.mark.django_db
@@ -263,10 +263,10 @@ def test_event_created_by_no_creds_private(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_no_creds(
+def test_event_private_created_by_no_creds(
         api_client,
         user_2,
-        event_not_visible_created_by_user_2,
+        event_private_created_by_user_2,
 ):
     response = api_client.get(
         reverse(CREATED_BY_URL, args=[user_2.id])
@@ -276,10 +276,10 @@ def test_event_not_visible_created_by_no_creds(
 
 
 @pytest.mark.django_db
-def test_event_not_visible_created_by_no_creds_private(
+def test_event_private_created_by_no_creds_private(
         api_client,
         user_2_private,
-        event_not_visible_created_by_user_2_private,
+        event_private_created_by_user_2_private,
 ):
     response = api_client.get(
         reverse(CREATED_BY_URL, args=[user_2_private.id])
