+++ apps/events/viewsets/tags.py
@@ -1,0 +1,34 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+
+from apps.events.models import Tag
+from apps.events.permissions import TagPermissions
+from apps.events.serializers import (
+    TagRetrieveSerializer,
+    TagCreateSerializer,
+    TagUpdateSerializer,
+    TagListSerializer,
+)
+
+
+class TagViewSet(viewsets.ModelViewSet):
+    """
+    A simple ViewSet for viewing and editing event Tags.
+    """
+
+    model = Tag
+    queryset = Tag.objects.all()
+    permission_classes = [IsAuthenticatedOrReadOnly, TagPermissions]
+    lookup_url_kwarg = "tag_id"
+    http_method_names = ["post", "get", "put", "delete"]
+
+    def get_serializer_class(self):
+        match self.action:
+            case "retrieve":
+                return TagRetrieveSerializer
+            case "create":
+                return TagCreateSerializer
+            case "update":
+                return TagUpdateSerializer
+            case "list":
+                return TagListSerializer
