+++ requirements/development.txt
@@ -1,1 +1,3 @@
 -r common.txt
+
+django-debug-toolbar==4.3.0
